refactor: reorganize public API
