+++ tests/archivey/test_open_compressed_stream.py
@@ -2,9 +2,9 @@ import logging
 
 import pytest
 
-from archivey.api.config import ArchiveyConfig
-from archivey.api.core import open_compressed_stream
-from archivey.api.exceptions import ArchiveNotSupportedError
+from archivey.config import ArchiveyConfig
+from archivey.core import open_compressed_stream
+from archivey.exceptions import ArchiveNotSupportedError
 from tests.archivey.sample_archives import SAMPLE_ARCHIVES, filter_archives
 from tests.archivey.testing_utils import skip_if_package_missing
