+++ tests/archivey/test_usage_errors.py
@@ -2,9 +2,9 @@ import logging
 
 import pytest
 
-from archivey.api.core import open_archive
-from archivey.api.exceptions import ArchiveMemberCannotBeOpenedError
-from archivey.api.types import ArchiveMember, MemberType
+from archivey.core import open_archive
+from archivey.exceptions import ArchiveMemberCannotBeOpenedError
+from archivey.types import ArchiveMember, MemberType
 from tests.archivey.sample_archives import (
     BASIC_ARCHIVES,
     SYMLINK_ARCHIVES,
